Add use of namespaces
Add explicit namespace use in class references for class that use a class from another area. This helps to identify a classes context when reading the unrelated class code.
ie. A document class that uses a Graphics class will also use Visuals::Graphics class name.

The only exception is the Messaging class as this simplifies the messaging notification calls that are used everywhere.